+++ Math.h
@@ -3,24 +3,24 @@
 
 #include "Vector.h"
 
-float squared_distance_between_points(const Vector* p1, const Vector* p2);
+double squared_distance_between_points(const Vector* p1, const Vector* p2);
 
-float cos_at_ab(float a2, float b2, float c2, float ab);
+double cos_at_ab(double a2, double b2, double c2, double ab);
 
-bool acute_angle(float cos);
+bool acute_angle(double cos);
 
 Vector cross_product(const Vector* a, const Vector* b, const Vector* c);
 
-float triangle_area(const Vector* a, const Vector* b, const Vector* c);
+double triangle_area(const Vector* a, const Vector* b, const Vector* c);
 
-bool segment_line_sphere_intersect(const Vector* p1, const Vector* p2, const Vector* s, float r);
+bool segment_line_sphere_intersect(const Vector* p1, const Vector* p2, const Vector* s, double r);
 
-float diffusion_coefficient(float temperature_K, float viscosity_eta, float diameter);
+double diffusion_coefficient(double temperature_K, double viscosity_eta, double diameter);
 
-float squared_displacement(int dimensions, float diffusion_coefficient, float tau);
+double squared_displacement(int dimensions, double diffusion_coefficient, double tau);
 
-float phi(float x);
+double phi(double x);
 
-float normal(float mean, float dev);
+double normal(double mean, double dev);
 
 #endif /* SIM_MATH_H */
