+++ RCell.h
@@ -13,7 +13,7 @@ class RCell : public Cell
 private:
 	std::vector<Molecule*>* rmolecules;
 public:
-	RCell(long identifier, Vector position, float radius);
+	RCell(long identifier, Vector position, double radius);
 	std::vector<Molecule*>* molecules();
 
 	bool collide(Molecule *m);
